+++ server/ModMetadata.cs
@@ -1,4 +1,4 @@
-﻿using SPTarkov.Server.Core.Models.Spt.Mod;
+using SPTarkov.Server.Core.Models.Spt.Mod;
 using SPTarkov.Server.Web;
 
 namespace RaidAdminPanel;
@@ -9,7 +9,7 @@ public record ModMetadata : AbstractModMetadata, IModWebMetadata
     public override string Name { get; init; } = "Raid Admin Panel";
     public override string Author { get; init; } = "kabzon93region";
     public override List<string>? Contributors { get; init; }
-    public override SemanticVersioning.Version Version { get; init; } = new("1.0.20");
+    public override SemanticVersioning.Version Version { get; init; } = new("1.0.21");
     public override SemanticVersioning.Range SptVersion { get; init; } = new("~4.0.0");
     public override List<string>? Incompatibilities { get; init; }
     public override Dictionary<string, SemanticVersioning.Range>? ModDependencies { get; init; }